Trait iced_style::pane_grid::StyleSheet

pub trait StyleSheet {
    fn picked_split(&self) -> Option<Line>;
    fn hovered_split(&self) -> Option<Line>;
}

A set of rules that dictate the style of a container.

Required Methods

fn picked_split(&self) -> Option<Line>

The Line to draw when a split is picked.

fn hovered_split(&self) -> Option<Line>

The Line to draw when a split is hovered.