Module iced_style::scrollable

Navigate an endless amount of content with a scrollbar.

Structs

Scrollbar

The appearance of a scrollable.

Scroller

The appearance of the scroller of a scrollable.

Traits

StyleSheet

A set of rules that dictate the style of a scrollable.