Trait iced_style::scrollable::StyleSheet

pub trait StyleSheet {
    type Style: Default + Copy;

    fn active(&self, style: Self::Style) -> Scrollbar;
    fn hovered(&self, style: Self::Style) -> Scrollbar;

    fn dragging(&self, style: Self::Style) -> Scrollbar { ... }
}

A set of rules that dictate the style of a scrollable.

Required Associated Types

type Style: Default + Copy

Required Methods

fn active(&self, style: Self::Style) -> Scrollbar

Produces the style of an active scrollbar.

fn hovered(&self, style: Self::Style) -> Scrollbar

Produces the style of an hovered scrollbar.

Provided Methods

fn dragging(&self, style: Self::Style) -> Scrollbar

Produces the style of a scrollbar that is being dragged.