Enum iced_style::theme::Theme

pub enum Theme {
    Light,
    Dark,
    Custom(Box<Custom>),
}

A built-in theme.

Variants

Light

The built-in light variant.

Dark

The built-in dark variant.

Custom(Box<Custom>)

A Theme that uses a Custom palette.

Implementations

impl Theme

pub const ALL: &'static [Self] = _

A list with all the defined themes.

pub fn custom(name: String, palette: Palette) -> Self

Creates a new custom Theme from the given Palette.

pub fn custom_with_fn( name: String, palette: Palette, generate: impl FnOnce(Palette) -> Extended, ) -> Self

Creates a new custom Theme from the given Palette, with a custom generator of a palette::Extended.

pub fn palette(&self) -> Palette

Returns the Palette of the Theme.

pub fn extended_palette(&self) -> &Extended

Returns the palette::Extended of the Theme.